Implement "--status-version 4" (JSON)
Parsing CSV can be cumbersome in depending software and prone to
breakage as has been obversed in the past when changing the CSV
format.

This patch allows the use of JSON as well using "status 4".

The "status 4" command uses the same keys/values as "status 2"
and "status 3", just formatted as JSON.